chore: prepare for new sensitive data layout
