+++ frontend/src/components/SensitiveData/index.ts
@@ -1,0 +1,3 @@
+import SensitiveColumnList from "./SensitiveColumnList.vue";
+
+export { SensitiveColumnList };
